+++ src/main/java/com/dumbdogdiner/betterwhitelist/commands/UnwhitelistCommand.java
@@ -1,8 +1,8 @@
-package com.dumbdogdiner.betterwhitelist_bungee.bungee.commands;
+package com.dumbdogdiner.betterwhitelist.commands;
 
-import com.dumbdogdiner.betterwhitelist_bungee.BaseClass;
-import com.dumbdogdiner.betterwhitelist_bungee.utils.MojangUser;
-import com.dumbdogdiner.betterwhitelist_bungee.utils.UsernameValidator;
+import com.dumbdogdiner.betterwhitelist.BaseClass;
+import com.dumbdogdiner.betterwhitelist.utils.MojangUser;
+import com.dumbdogdiner.betterwhitelist.utils.UsernameValidator;
 
 import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.CommandSender;
