+++ src/main/java/com/dumbdogdiner/betterwhitelist/discord/commands/HelpCommand.java
@@ -1,8 +1,8 @@
-package com.dumbdogdiner.betterwhitelist_bungee.discord.commands;
+package com.dumbdogdiner.betterwhitelist.discord.commands;
 
-import com.dumbdogdiner.betterwhitelist_bungee.BaseClass;
-import com.dumbdogdiner.betterwhitelist_bungee.BetterWhitelistBungee;
-import com.dumbdogdiner.betterwhitelist_bungee.discord.lib.Command;
+import com.dumbdogdiner.betterwhitelist.BaseClass;
+import com.dumbdogdiner.betterwhitelist.BetterWhitelistBungee;
+import com.dumbdogdiner.betterwhitelist.discord.lib.Command;
 
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
