+++ src/main/java/com/dumbdogdiner/betterwhitelist/discord/commands/UnwhitelistCommand.java
@@ -1,8 +1,8 @@
-package com.dumbdogdiner.betterwhitelist_bungee.discord.commands;
+package com.dumbdogdiner.betterwhitelist.discord.commands;
 
-import com.dumbdogdiner.betterwhitelist_bungee.BaseClass;
-import com.dumbdogdiner.betterwhitelist_bungee.discord.lib.Command;
-import com.dumbdogdiner.betterwhitelist_bungee.discord.utils.RoleUtil;
+import com.dumbdogdiner.betterwhitelist.BaseClass;
+import com.dumbdogdiner.betterwhitelist.discord.lib.Command;
+import com.dumbdogdiner.betterwhitelist.discord.utils.RoleUtil;
 
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
