+++ src/main/java/com/dumbdogdiner/betterwhitelist/discord/commands/WhitelistCommand.java
@@ -1,10 +1,10 @@
-package com.dumbdogdiner.betterwhitelist_bungee.discord.commands;
+package com.dumbdogdiner.betterwhitelist.discord.commands;
 
-import com.dumbdogdiner.betterwhitelist_bungee.BaseClass;
-import com.dumbdogdiner.betterwhitelist_bungee.discord.lib.Command;
-import com.dumbdogdiner.betterwhitelist_bungee.discord.utils.RoleUtil;
-import com.dumbdogdiner.betterwhitelist_bungee.utils.MojangUser;
-import com.dumbdogdiner.betterwhitelist_bungee.utils.UsernameValidator;
+import com.dumbdogdiner.betterwhitelist.BaseClass;
+import com.dumbdogdiner.betterwhitelist.discord.lib.Command;
+import com.dumbdogdiner.betterwhitelist.discord.utils.RoleUtil;
+import com.dumbdogdiner.betterwhitelist.utils.MojangUser;
+import com.dumbdogdiner.betterwhitelist.utils.UsernameValidator;
 
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
