+++ src/main/java/com/dumbdogdiner/betterwhitelist/discord/listeners/ReadyListener.java
@@ -1,4 +1,4 @@
-package com.dumbdogdiner.betterwhitelist_bungee.discord.listeners;
+package com.dumbdogdiner.betterwhitelist.discord.listeners;
 
 import net.dv8tion.jda.api.events.ReadyEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
