+++ src/main/java/com/dumbdogdiner/betterwhitelist/discord/utils/RoleUtil.java
@@ -1,6 +1,6 @@
-package com.dumbdogdiner.betterwhitelist_bungee.discord.utils;
+package com.dumbdogdiner.betterwhitelist.discord.utils;
 
-import com.dumbdogdiner.betterwhitelist_bungee.BaseClass;
+import com.dumbdogdiner.betterwhitelist.BaseClass;
 
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.Role;
