+++ src/main/resources/plugin.yml
@@ -1,4 +1,4 @@
 name: BetterWhitelistBungee
 version: ${project.version}
 author: 'kokumaji, fuzzyfoxie, vladfrangu, Waffaru'
-main: com.dumbdogdiner.betterwhitelist_bungee.BetterWhitelistBungee
+main: com.dumbdogdiner.betterwhitelist.BetterWhitelistBungee
